fixup migration and seeder

+++ database/migrations/2023_09_13_180301_add_location_id_to_autoparts.php
@@ -9,7 +9,7 @@ class AddLocationIdToAutoparts extends Migration
     public function up()
     {
         Schema::table('autoparts', function (Blueprint $table) {
-            $table->unsignedInteger('location_id')->nullable()->after('description');
+            $table->unsignedBigInteger('location_id')->nullable()->after('description');
 
             $table->foreign('location_id')->references('id')->on('autopart_list_locations');
         });
